requirements.txt
customtkinter
cryptography
pillow
pyperclip